Light refactoring for consistent API layout
