+++ src/murfey/server/api/display.py
