+++ src/murfey/server/api/instrument.py
